fix: long indexing migration batch 3 - H2, H13 fixed
Fixed HIGH priority issues:
- H1: Confirmed already fixed (all typed getters have long[] overloads)
- H2: Added missing long[] overloads for 9 typed setters in NDArray:
  SetBoolean, SetByte, SetInt16, SetUInt16, SetUInt32, SetUInt64,
  SetChar, SetSingle, SetDecimal
- H13: ArgMax/ArgMin now return Int64 indices (supports >2B arrays):
  - ReductionKernel.cs: ResultType returns NPTypeCode.Int64 for ArgMax/ArgMin
  - ArgMaxSimdHelper/ArgMinSimdHelper: return long, take long totalSize
  - All loop counters/indices changed from int to long

Total fixed this session: 15 issues
Remaining HIGH: H5, H7 (both acceptable - protected by .NET limits)

+++ src/NumSharp.Core/Backends/Kernels/ReductionKernel.cs
@@ -36,15 +36,15 @@ namespace NumSharp.Backends.Kernels
 
         /// <summary>
         /// Result type depends on operation:
-        /// - ArgMax/ArgMin: always Int32
+        /// - ArgMax/ArgMin: always Int64 (supports >2B element arrays)
         /// - All/Any: always Boolean
         /// - Mean: always Double (or accumulator type if specified)
         /// - Others: same as accumulator type
         /// </summary>
         public NPTypeCode ResultType => Op switch
         {
-            ReductionOp.ArgMax => NPTypeCode.Int32,
-            ReductionOp.ArgMin => NPTypeCode.Int32,
+            ReductionOp.ArgMax => NPTypeCode.Int64,
+            ReductionOp.ArgMin => NPTypeCode.Int64,
             ReductionOp.All => NPTypeCode.Boolean,
             ReductionOp.Any => NPTypeCode.Boolean,
             _ => AccumulatorType
@@ -76,11 +76,13 @@ namespace NumSharp.Backends.Kernels
 
         /// <summary>
         /// Result type depends on operation.
+        /// - ArgMax/ArgMin: always Int64 (supports >2B element arrays)
+        /// - Others: same as accumulator type
         /// </summary>
         public NPTypeCode ResultType => Op switch
         {
-            ReductionOp.ArgMax => NPTypeCode.Int32,
-            ReductionOp.ArgMin => NPTypeCode.Int32,
+            ReductionOp.ArgMax => NPTypeCode.Int64,
+            ReductionOp.ArgMin => NPTypeCode.Int64,
             _ => AccumulatorType
         };
